Huy hieu feature

+++ Entities/DeviceDriver/ThresholdDeviceEntity.cs
@@ -12,8 +12,8 @@ namespace Entities
         public string TypeDevice { get; set; } = null!;
         public bool? OnInUpperThreshold { get; set; }
         // Case: Device return value
-        public double? ThresholdValueOn { get; set; }
-        public double? ThresholdValueOff { get; set; }
+        public int? ThresholdValueOn { get; set; }
+        public int? ThresholdValueOff { get; set; }
         public bool IsDelete { get; set; } = false;
     }
 }
